% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceToReferencePlane.R
\name{distanceToReferencePlane}
\alias{distanceToReferencePlane}
\title{Score based on the distance to a healthy/reference plane}
\arguments{
\item{x}{A phyloseq object}

\item{dist_mat}{A distance matrix. Can be output of \code{phyloseq::distance}
or \code{vegan::vegdist}.}

\item{reference_samples}{Vector of samples to use as reference.}
}
\value{
A data frame with distance to reference plane ('dtrpScore') values and sample information.
}
\description{
Score based on the distance to a healthy/reference plane
}
\details{
Calculates a 'healthy' or 'reference plane'. The plane is calculated in a space
         derived from PCoA and based on user-defined distances between samples from healthy
         subjects. Briefly, a model is constructed using the samples from healthy subjects,
         which were fitted to a two-dimensional plane embedded in a three-dimensional space using
         the least-squares method. The plane is then restricted to only span the three-dimensional
         ranges of the healthy control samples. The plane is considered a proxy for the normal
         microbial variation in healthy subjects. \code{distanceToReferencePlane} calculates the
         Euclidean distance (within PCoA space) from each sample towards the plane, which can
         be considered a measure of (ab)normality. The user can provide a custom distance matrix.
         If the user provides a UniFrac distance matrix, then the resulting score is comparable to
         the \code{dysbiosis score} reported in \code{Halfvarson J, Brislawn CJ, Lamendella R et al.
         (2017)}.
}
\examples{
library(dysbiosisR)
# We use WirbelJ_2018 as test data
dist.mat <- phyloseq::distance(WirbelJ_2018, "bray")
ref.samples <- sample_names(subset_samples(WirbelJ_2018,
                                           disease == "healthy"))
dtrp.results <- distanceToReferencePlane(WirbelJ_2018,
                                         dist_mat = dist.mat,
                                         reference_samples = ref.samples)

head(dtrp.results)

}
\references{
\itemize{
\item{}{Halfvarson J, Brislawn CJ, Lamendella R et al. (2017)
Dynamics of the human gut microbiome in inflammatory bowel disease.
\emph{Nature Microbiology}, 2, article number: 17004.}

}
}
\author{
Wouter A.A. de Steenhuijsen Piters
}
