% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dysbiosisMedianCLV.R
\name{dysbiosisMedianCLV}
\alias{dysbiosisMedianCLV}
\title{Dysbiosis Score Based on Median Community Level Variation}
\arguments{
\item{x}{A phyloseq object}

\item{dist_mat}{A distance matrix. Can be output of \code{phyloseq::distance}
or \code{vegan::vegdist}.}

\item{reference_samples}{Vector of samples to use as reference.}
}
\value{
A data frame with Median CLV values and sample information
}
\description{
Dysbiosis Score Based on Median Community Level Variation
}
\details{
Calculates median variation in a given sample compared to a reference
         sample group. Here \code{dysbiosisMedianCLV} will calculate median
         variation for a sample compared to a reference sample set. The user
         can provide a custom distance matrix. If the user provides a Bray-Curtis
         dissimilarity matrix, then the resulting score is comparable to the
         \code{dysbiosis score} reported in
         \code{Lloyd-Price J, Arze C, Ananthakrishnan AN et al. (2019)}.
}
\examples{
library(dysbiosisR)
# We use WirbelJ_2018 as test data
dist.mat <- phyloseq::distance(WirbelJ_2018, "bray")
ref.samples <- rownames(meta(subset_samples(WirbelJ_2018, disease == "healthy")))
db.1 <- dysbiosisMedianCLV(WirbelJ_2018,
                           dist_mat = dist.mat,
                           reference_samples = ref.samples)

head(db.1)

}
\references{
\itemize{
\item{}{Lloyd-Price J, Arze C, Ananthakrishnan AN et al. (2019).
Multi-omics of the gut microbial ecosystem in inflammatory bowel diseases.
\emph{Nature}, 569(7758), pp.655-662.}

}
}
\author{
Sudarshan A. Shetty
}
