% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dysbiosisOBB.R
\name{dysbiosisOBB}
\alias{dysbiosisOBB}
\title{Random Forest Prediction Based Score}
\arguments{
\item{x}{A phyloseq object}

\item{group_col}{A column in \code{phyloseq::sample_data} with all control and
case labels.}

\item{case_label}{A character string specifying case/disease labels in
group_col.}

\item{seed_value}{The random seed of your session for reproducibility.}

\item{add_tuneRF_params}{A list of arguments to pass for
\code{randomForest::tuneRF}}

\item{ntree}{Number of trees. See \code{randomForest::randomForest}}

\item{plot_roc}{Logical TRUE or FALSE to plot ROC curve.}

\item{...}{Additional arguments if necessary for
\code{randomForest::randomForest}.}
}
\value{
A data frame with oob.score score and sample information
}
\description{
Random Forest Prediction Based Score
}
\details{
An out-of-bag probability of Random Forest of being classified as
         case or diseased group.

The original article \code{Saffouri GB, Shields-Cutler R et al. 2019}
reported a Symptom Index abbreviated as SI. In this approach the feature
abundances are used for \code{Random Forest} classification and the resulting
out of bag (OOB) predicted probability of being classified in disease group
is considered as an SI or also dysbiosis index. The \code{dysbiosisOBB}
function in this package allows for calculating this measure with some level
of freedom of 'tuneRF' and 'randomForest' parameters via the \code{randomForest}
R package.
}
\examples{
data("WirbelJ_2018")
# data are relative abundances summed to 100 or
# near 100 val
dysbiosis.oob <- dysbiosisOBB(WirbelJ_2018,
                              group_col = "disease",
                              control_label = "healthy",
                              case_label = "CRC",
                              seed_value = 1235,
                              add_tuneRF_params = list(ntreeTry=100,
                                                       stepFactor=1.5,
                                                       improve=0.01,
                                                       trace=TRUE,
                                                       dobest=FALSE),
                              ntree = 100, # increase for real data
                              plot_roc = TRUE)
head(dysbiosis.oob)

}
\references{
\itemize{
\item{}{Saffouri GB, Shields-Cutler R et al. (2019).
Small intestinal microbial dysbiosis underlies symptoms associated with
functional gastrointestinal disorders.
\emph{Nature communications}, 10(1), pp.1-11.}
}
}
\author{
Sudarshan A. Shetty
}
