% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucledianDistCentroids.R
\name{euclideanDistCentroids}
\alias{euclideanDistCentroids}
\title{Dysbiosis Score Based on Euclidean Distance to Group Centroids}
\arguments{
\item{x}{A phyloseq object}

\item{dist_mat}{A distance matrix. Can be output of \code{phyloseq::distance}
or \code{vegan::vegdist}.}

\item{use_squared}{Logical. Default is FALSE. If TURE to the score is
calculated using the squared distance to group centroids.
see \code{usedist::dist_to_centroids}.}

\item{group_col}{A column in \code{phyloseq::sample_data} with all control and
case labels.}

\item{control_label}{A character string specifying control/healthy labels in
group_col.}

\item{case_label}{A character string specifying case/disease labels in
group_col.}
}
\value{
A data frame with Centroid distance to each group and a score with
        sample information
}
\description{
Dysbiosis Score Based on Euclidean Distance to Group Centroids
}
\details{
Calculates difference in \code{euclidean distance (ED)} for a sample
         to group centroids. For example, sample_1 to control centroid minus
         sample_1 to case centroid. The user can provide a custom distance
         matrix. This approach was used in \code{AlShawaqfeh MK et al. (2017)}.
}
\examples{
library(dysbiosisR)
# We use WirbelJ_2018 as test data
dist.mat <- phyloseq::distance(WirbelJ_2018, "bray")
db.1 <- euclideanDistCentroids(WirbelJ_2018,
                               dist_mat = dist.mat,
                               use_squared = TRUE,
                               group_col = "disease",
                               control_label = "healthy",
                               case_label = "CRC")

head(db.1)

}
\references{
\itemize{
\item{}{AlShawaqfeh MK et al. (2017).
A dysbiosis index to assess microbial changes in fecal samples of dogs with
chronic inflammatory enteropathy.
\emph{FEMS microbiology ecology}, 93(11), p.fix136.}

}
}
\author{
Sudarshan A. Shetty
}
