% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDysbiosis.R
\name{plotDysbiosis}
\alias{plotDysbiosis}
\title{Plot Dysbiosis}
\arguments{
\item{df}{The data frame output from different dysbiosis calculators provided
by dysbiosisR package.}

\item{xvar}{Variable to plot on x-axis.}

\item{yvar}{Variable to plot on y-axis.}

\item{colors}{colors to use for plotting. Default is NULL.}
}
\value{
A ggplot2 object
}
\description{
Plot Dysbiosis
}
\details{
A wrapper for quick plot.
}
\examples{
data("WirbelJ_2018")
# data are relative abundances summed to 100 or
# near 100 val
dysbiosis.oob <- dysbiosisOBB(WirbelJ_2018,
                              group_col = "disease",
                              control_label = "healthy",
                              case_label = "CRC",
                              seed_value = 1235,
                              add_tuneRF_params = list(ntreeTry=100,
                                                       stepFactor=1.5,
                                                       improve=0.01,
                                                       trace=TRUE,
                                                       dobest=FALSE),
                              ntree = 100, # increase for real data
                              plot_roc = TRUE)
plotDysbiosis(df=dysbiosis.oob,
              xvar="disease",
              yvar="oob.score",
              colors=c(CRC="brown3", healthy="steelblue"))
}
\author{
Sudarshan A. Shetty
}
