% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudStatistic.R
\name{cloudStatistic}
\alias{cloudStatistic}
\title{Cloud-based LOcally linear Unbiased Dysbiosis (CLOUD) test}
\arguments{
\item{x}{A phyloseq object}

\item{reference_samples}{Vector of samples to use as reference.}

\item{dist_mat}{A distance matrix. Can be output of \code{phyloseq::distance}
or \code{vegan::vegdist}.}

\item{k_num}{Neighbors to use. Default is 80 percent. User can
define percent of samples. See \code{Montassier E et al. 2018}
for more details.}

\item{ndim}{Dimension of the space in which the data are to be represented.
Default is -1. See \code{Montassier E et al. 2018}}
}
\value{
A data frame with CLOUD stats and sample information
}
\description{
Cloud-based LOcally linear Unbiased Dysbiosis (CLOUD) test
}
\details{
Calculates CLOUD score.

Cloud-based LOcally linear Unbiased Dysbiosis (CLOUD) test is a non-parametric
test and returns a measure of dysbiosis. The function was adapted from
the original article by \code{Montassier E et al. 2018}. Here, a user defines
a set of reference samples from which distance of every other sample is
calculated. When calculating the \code{CLOUD} stats the \emph{k} is an
important parameter specified by argument \code{k_num}. By default we use a
conservative 80 percent of the samples in each group.
}
\examples{
data("WirbelJ_2018")
library(phyloseq)
ps <- WirbelJ_2018
# Define controls as reference samples
ref.samples <- sample_names(subset_samples(WirbelJ_2018,
                                           disease == "healthy"))
dist.data <- phyloseq::distance(ps, "bray")
cloud.results <- cloudStatistic(x= ps,
                                dist_mat = dist.data,
                                reference_samples = ref.samples,
                                ndim=-1)
head(cloud.results)

}
\references{
\itemize{
\item{}{Lloyd-Price J, Arze C, Ananthakrishnan AN et al. (2019).
Multi-omics of the gut microbial ecosystem in inflammatory bowel diseases.
\emph{Nature}, 569(7758), pp.655-662.}

\item{}{Montassier E et al. (2018). CLOUD: a non-parametric detection test
for microbiome outliers.
\emph{Microbiome}, 6(1), pp.1-10.}

}
}
\author{
Sudarshan A. Shetty
}
